feat: the upper level of hrl (#64)
* Create sb3.py

* Update sb3.py

* feat: cpy sb3 dqn

* Update sb3.py

* rm save model

* Update dqn.py

* feat: init Q learning

* Update dqn.py

* Update sb3.py

* feat: add. a wrapper

* Update sb3.py

* Update sb3.py

* Update sb3.py

* feat: change model neurons

* feat: change buffer

* Update sb3.py

* Update sb3.py

* Update sb3.py

* Update sb3.py

* feat: replay_buffer

* update trajectory buffer

* Update sb3.py

* Update sb3.py

* Update fmp.py

* add buffer repr

* Update sb3.py

* feat: finish buffer

* fix: sample a traj rather than trans

* fix: dead loop

* feat: add verbose for training

* feat: use per buffer

* embed pruned

* to train

* feat: dqn draft based on the assumption that some apis are defined

* dqn

* fix q network

* Update dqn.py

* fix tolist

* remove print

* resolve buffer choice

* fix: targets

* small fix

* Update sb3.py

* Revert "Update sb3.py"

This reverts commit 163e0d26bbcd0f9194ee40d3ba23cf63daf268d6.

* Update sb3.py

* change buffer prob

* Update sb3.py

* Update sb3.py

* handle done and charge case

* Update sb3.py

* add loss metrics

* Update sb3.py

* modify reward for charging & embed metrics

* add tbc metric

* safe_indicator

* embed safe_indicator

* refactor: reformat

* remove location in state

* Update fmp.py

* modity status indicator

* handle no move status

* small fix

* feat: gradual reward

* Update fmp-numerical.py

* Update fmp.py

* Update fmp.py

* Update fmp.py

* enhance safe indicator

* feat: update reward addition according to the safe indicator

* fix safe_indicator

* safe indicator

* finish step_move

* Update fmp.py

* finish except reward

* feat: fix some small bugs

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* draft

* fix stopping issue

* fix no resume issue

* fix render issue

* Hrl mdp (#63)

* feat: hrl - mdp

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* buffer problem

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* feat:

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* fix: buffer

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update sb3.py

* Update sb3.py

* runnable

* restore to default param

* Update dqn.py

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci


* Update dqn.py

* add no render option

* Update sb3.py

* Update sb3.py

* Update sb3.py

* Update dqn.py

* Update sb3.py

* update

* Revert "update"

This reverts commit 972c98f68442455bb4c4fe686e55184141c15be8.

* fine tune the params

* I think it;'s okay for now

* dummy sample

* fix batch size problem

* Update dqn.py

* remove json

* train json

* Update sb3.py

* Create q-learning.py

* add defined replay buffer

* Create sb3-debug.py

* loss reward real time and plot

* plot

* fix

* smooth curve

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update fmp.py

+++ src/sumo_gym/utils/xml_utils.py
@@ -35,7 +35,7 @@ def get_electric_vehicles(flow_xml_tree):
     maxBatteryCap = -1
     for vtype_param in vtype_params:
         if vtype_param.attrib["key"] == "maximumBatteryCapacity":
-            maxBatteryCap = float(vtype_param.attrib["value"])
+            maxBatteryCap = int(vtype_param.attrib["value"])
 
     ev_lst = []
     vehicles = flow_xml_tree.findall(VEHICLE_XML_TAG)
